Fix Intel CC warning.

+++ src/sbbs3/xmodem.c
@@ -447,7 +447,7 @@ BOOL xmodem_send_file(xmodem_t* xm, const char* fname, FILE* fp, time_t* start, 
 			memset(block,0,sizeof(block));
 			SAFECOPY(block,getfname(fname));
 			i=sprintf(block+strlen(block)+1,"%lu %lo 0 0 %d %ld"
-				,st.st_size
+				,(ulong)st.st_size
 				,st.st_mtime
 				,xm->total_files-xm->sent_files
 				,xm->total_bytes-xm->sent_bytes);
